Big entity update!!!
This is the start of an update to allow for entity creation in blender.
Tutorial will need to be updated, and lots more structs need to be
added.
A 'utilities' folder has been added in nms_imp which contains a little
script to create blank structs. Might make a more powerful one later.