Install and run Pytest from within Blender
Here's a basic implementation of Pytest in Blender. This adds a new panel, and checks for Pytest. If it's not there, it installs Pytest from Pip (Included in 2.81+)

Right now the tests need to be listed in pytest.py.
This needs an add-on option to disable the tests panel

NMS gets updated frequently and breaks frequently so this makes testing available to users